# swagger_server/models/check_token_assignment_output.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class CheckTokenAssignmentOutput(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, token: str=None, status: str=None):  # noqa: E501
        """CheckTokenAssignmentOutput - a model defined in Swagger

        :param token: The token of this CheckTokenAssignmentOutput.  # noqa: E501
        :type token: str
        :param status: The status of this CheckTokenAssignmentOutput.  # noqa: E501
        :type status: str
        """
        self.swagger_types = {
            'token': str,
            'status': str
        }

        self.attribute_map = {
            'token': 'token',
            'status': 'status'
        }
        self._token = token
        self._status = status

    @classmethod
    def from_dict(cls, dikt) -> 'CheckTokenAssignmentOutput':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CheckTokenAssignmentOutput of this CheckTokenAssignmentOutput.  # noqa: E501
        :rtype: CheckTokenAssignmentOutput
        """
        return util.deserialize_model(dikt, cls)

    @property
    def token(self) -> str:
        """Gets the token of this CheckTokenAssignmentOutput.

        The token assigned to the counter  # noqa: E501

        :return: The token of this CheckTokenAssignmentOutput.
        :rtype: str
        """
        return self._token

    @token.setter
    def token(self, token: str):
        """Sets the token of this CheckTokenAssignmentOutput.

        The token assigned to the counter  # noqa: E501

        :param token: The token of this CheckTokenAssignmentOutput.
        :type token: str
        """
        if token is None:
            raise ValueError("Invalid value for `token`, must not be `None`")  # noqa: E501

        self._token = token

    @property
    def status(self) -> str:
        """Gets the status of this CheckTokenAssignmentOutput.

        The status of the counter  # noqa: E501

        :return: The status of this CheckTokenAssignmentOutput.
        :rtype: str
        """
        return self._status

    @status.setter
    def status(self, status: str):
        """Sets the status of this CheckTokenAssignmentOutput.

        The status of the counter  # noqa: E501

        :param status: The status of this CheckTokenAssignmentOutput.
        :type status: str
        """
        if status is None:
            raise ValueError("Invalid value for `status`, must not be `None`")  # noqa: E501

        self._status = status