# swagger_server/models/generate_token_input.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class GenerateTokenInput(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, service_id: str=None):  # noqa: E501
        """GenerateTokenInput - a model defined in Swagger

        :param service_id: The service_id of this GenerateTokenInput.  # noqa: E501
        :type service_id: str
        """
        self.swagger_types = {
            'service_id': str
        }

        self.attribute_map = {
            'service_id': 'service_id'
        }
        self._service_id = service_id

    @classmethod
    def from_dict(cls, dikt) -> 'GenerateTokenInput':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The GenerateTokenInput of this GenerateTokenInput.  # noqa: E501
        :rtype: GenerateTokenInput
        """
        return util.deserialize_model(dikt, cls)

    @property
    def service_id(self) -> str:
        """Gets the service_id of this GenerateTokenInput.

        ID of the service for which token needs to be generated  # noqa: E501

        :return: The service_id of this GenerateTokenInput.
        :rtype: str
        """
        return self._service_id

    @service_id.setter
    def service_id(self, service_id: str):
        """Sets the service_id of this GenerateTokenInput.

        ID of the service for which token needs to be generated  # noqa: E501

        :param service_id: The service_id of this GenerateTokenInput.
        :type service_id: str
        """
        if service_id is None:
            raise ValueError("Invalid value for `service_id`, must not be `None`")  # noqa: E501

        self._service_id = service_id